requirements.txt
dnspython
boto3
pyyaml
pyopenssl
